Fix listing of queries in query edit view; fixes leave in/out.
The problem is that a filter applied to the query list doesn't work.  All
queries are returned no matter what the value passed in for private_for is.
So I added conditions instead.  Note that the condition I added is more
accurate: we want public queries the user owns to be listed in the first
section where they can make it private if they want to, and we don't want to
list retired queries in either that or the public queries section.

The reason there was a permission error when trying to remove or add a query
is that when the form was submitted it tried to update properties on queries
that the submitting user didn't own, because the form was constructed wrong
(those properties should only have been set on queries the user does own).