Update GitHub email visibility failure message
The previous link didn't contain the option necessary to make the email public. While the email settings page may need to be edited as well, the current github UI contains a pointer from the newly linked "Public email" section to the email settings page so users should be better guided to do the right update.