Update guess_version function for traceback links
s/master/main/, and add support for double-digit minor versions.

This also removes the check for specific version numbers.  This may open
us up to creating some bad links, (2.8, anyone?) but chances are that
any link we make in such cases would be a bad one anyway.  We are still
restricted to 2.x and 3.x, though.

Closes python/bugs.python.org#61